Split AndroidNotificationCenter.cs file into several files like AndroidNotification.cs, AndroidNotificationChannel.cs etc.
