+++ com.unity.mobile.notifications/Runtime/Android/AndroidNotification.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5920d00d82f60ad438e2d483048328d6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
