+++ com.unity.mobile.notifications/Runtime/Android/AndroidNotificationChannel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f3829a6c0515ee45b4e641fbd0999f6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
