+++ com.unity.mobile.notifications/Runtime/Android/AndroidNotificationIntentData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25b01cf50533ff64387fe382f4075a2c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
